feature (resource): put v2 custom custom converters in central place
