+++ src/models/v2/ontologies/standoff-class-definition.ts
@@ -1,6 +1,7 @@
 import {JsonObject, JsonProperty} from 'json2typescript';
 import {Constants} from '../Constants';
-import {ClassDefinition, IHasProperty, PropertiesListConverter, SubClassOfConverter} from './class-definition';
+import {ClassDefinition, IHasProperty} from './class-definition';
+import {PropertiesListConverter, SubClassOfConverter} from '../CustomConverters';
 
 @JsonObject('StandoffClassDefinition')
 export class StandoffClassDefinition extends ClassDefinition {
